requirements list.txt
absl-py                       1.4.0
altair                        5.1.1
apache-beam                   2.46.0
array-record                  0.2.0
asttokens                     2.2.1
astunparse                    1.6.3
attrs                         23.1.0
avro-python3                  1.10.2
backcall                      0.2.0
blinker                       1.6.2
cached-property               1.5.2
cachetools                    5.3.0
certifi                       2022.12.7
charset-normalizer            3.1.0
chex                          0.1.7
click                         8.1.3
cloudpickle                   2.2.1
colorama                      0.4.6
comm                          0.1.3
contextlib2                   21.6.0
contourpy                     1.0.7
crcmod                        1.7
cycler                        0.11.0
Cython                        0.29.34
debugpy                       1.6.7
decorator                     5.1.1
dill                          0.3.1.1
dm-tree                       0.1.8
docopt                        0.6.2
etils                         1.2.0
executing                     1.2.0
fastavro                      1.7.3
fasteners                     0.18
flatbuffers                   2.0.7
flax                          0.6.10
fonttools                     4.39.3
gast                          0.4.0
gin-config                    0.5.0
gitdb                         4.0.10
GitPython                     3.1.37
google-api-core               2.11.0
google-api-python-client      2.85.0
google-auth                   2.17.3
google-auth-httplib2          0.1.0
google-auth-oauthlib          0.4.6
google-pasta                  0.2.0
googleapis-common-protos      1.59.0
grpcio                        1.53.0
h5py                          3.8.0
hdfs                          2.7.0
httplib2                      0.21.0
idna                          3.4
importlib-metadata            6.8.0
importlib-resources           5.12.0
ipykernel                     6.22.0
ipython                       8.12.0
jax                           0.4.16
jaxlib                        0.4.14
jedi                          0.18.2
Jinja2                        3.1.2
jsonschema                    4.19.1
jsonschema-specifications     2023.7.1
jupyter_client                8.1.0
jupyter_core                  5.3.0
kaggle                        1.5.13
keras                         2.10.0
Keras-Preprocessing           1.1.2
kiwisolver                    1.4.4
libclang                      16.0.0
lvis                          0.5.3
lxml                          4.9.2
Markdown                      3.4.3
markdown-it-py                2.2.0
MarkupSafe                    2.1.2
matplotlib                    3.7.1
matplotlib-inline             0.1.6
mdurl                         0.1.2
ml-dtypes                     0.3.0
msgpack                       1.0.5
nest-asyncio                  1.5.6
numpy                         1.23.5
oauthlib                      3.2.2
object-detection              0.1
objsize                       0.6.1
opencv-python                 4.7.0.72
opencv-python-headless        4.8.0.76
opt-einsum                    3.3.0
optax                         0.1.5
orbax-checkpoint              0.2.3
orjson                        3.8.10
packaging                     20.9
pandas                        2.0.0
parso                         0.8.3
pickleshare                   0.7.5
Pillow                        9.5.0
pip                           23.1.2
platformdirs                  3.2.0
portalocker                   2.8.2
promise                       2.3
prompt-toolkit                3.0.38
proto-plus                    1.22.2
protobuf                      3.20.2
psutil                        5.9.4
pure-eval                     0.2.2
pyarrow                       9.0.0
pyasn1                        0.4.8
pyasn1-modules                0.2.8
pycocotools                   2.0.6
pydeck                        0.8.1b0
pydot                         1.4.2
Pygments                      2.14.0
pymongo                       3.13.0
pyparsing                     2.4.7
PyQt5                         5.15.9
PyQt5-Qt5                     5.15.2
PyQt5-sip                     12.11.1
python-dateutil               2.8.2
python-slugify                8.0.1
python-version                0.0.2
pytz                          2023.3
pywin32                       306
PyYAML                        6.0
pyzmq                         25.0.2
referencing                   0.30.2
regex                         2023.3.23
requests                      2.28.2
requests-oauthlib             1.3.1
rich                          13.3.5
rpds-py                       0.10.3
rsa                           4.9
sacrebleu                     2.2.0
scipy                         1.10.1
setuptools                    67.7.1
six                           1.16.0
slim                          0.1       c:\users\owner\desktop\machine learning\tfodcourse\tensorflow\models\research\slim
smmap                         5.0.1
stack-data                    0.6.2
streamlit                     1.27.0
tabulate                      0.9.0
tenacity                      8.2.3
tensorboard                   2.10.1
tensorboard-data-server       0.6.1
tensorboard-plugin-wit        1.8.1
tensorflow                    2.10.1
tensorflow-addons             0.20.0
tensorflow-datasets           4.9.2
tensorflow-estimator          2.10.0
tensorflow-hub                0.12.0
tensorflow-intel              2.12.0
tensorflow-io                 0.31.0
tensorflow-io-gcs-filesystem  0.31.0
tensorflow-metadata           1.13.1
tensorflow-model-optimization 0.7.4
tensorflow-text               2.10.0
tensorflowjs                  3.21.0
tensorstore                   0.1.36
termcolor                     2.2.0
text-unidecode                1.3
tf-models-official            2.12.0
tf-slim                       1.1.0
tk                            0.1.0
toml                          0.10.2
toolz                         0.12.0
tornado                       6.2
tqdm                          4.65.0
traitlets                     5.9.0
typeguard                     2.13.3
typing_extensions             4.5.0
tzdata                        2023.3
tzlocal                       5.0.1
uritemplate                   4.1.1
urllib3                       1.26.15
validators                    0.22.0
versioneer                    0.28
watchdog                      3.0.0
wcwidth                       0.2.6
Werkzeug                      2.2.3
wget                          3.2
wheel                         0.40.0
wrapt                         1.14.1
zipp                          3.15.0
zstandard                     0.20.0